requirements.txt
# DH requirements
deephaven-plugin-plotly-express
deephaven-plugin-ui
